chore(suite): reorg Preloader dir
