+++ packages/suite/src/components/suite/AppNavigation.tsx
@@ -1,16 +1,13 @@
 import { useRef, useLayoutEffect, useState, ReactNode } from 'react';
-
 import styled, { css } from 'styled-components';
 
 import { variables, IconProps, useTheme, Button, Icon, Dropdown } from '@trezor/components';
 import { Route } from '@suite-common/suite-types';
 
-import { AccountFormCloseButton, AppNavigationTooltip } from 'src/components/suite';
+import { AccountFormCloseButton, AppNavigationTooltip, HoverAnimation } from 'src/components/suite';
 import { useSelector } from 'src/hooks/suite';
 import { MAX_WIDTH, MAX_WIDTH_WALLET_CONTENT } from 'src/constants/suite/layout';
-
 import { AccountStickyContent } from './AccountStickyContent';
-import { HoverAnimation } from './NavigationBar/components/HoverAnimation';
 
 const { FONT_WEIGHT, FONT_SIZE } = variables;
