+++ packages/suite/src/components/suite/Preloader/SuiteLayout/NavigationBar/DeviceSelector/DeviceStatus.tsx
