+++ packages/suite/src/components/suite/Preloader/SuiteLayout/NavigationBar/MainNavigation.tsx
@@ -3,11 +3,10 @@ import styled, { css } from 'styled-components';
 
 import { findRouteByName } from 'src/utils/suite/router';
 import { variables } from '@trezor/components';
-import { Translation } from 'src/components/suite';
+import { HoverAnimation, Translation } from 'src/components/suite';
 import { MAIN_MENU_ITEMS } from 'src/constants/suite/menu';
 import { useAccountSearch, useSelector, useDispatch } from 'src/hooks/suite';
 import { goto } from 'src/actions/suite/routerActions';
-import { HoverAnimation } from './HoverAnimation';
 
 interface ComponentProps {
     isActive: boolean;
