+++ packages/suite/src/components/suite/Preloader/SuiteLayout/NavigationBar/NavigationActions/ActionItem.tsx
@@ -1,8 +1,10 @@
 import { useMemo, Ref, forwardRef, ReactNode, HTMLAttributes } from 'react';
 import styled from 'styled-components';
+
 import { useTheme, Icon, IconProps, variables, Spinner, Tooltip } from '@trezor/components';
 import { FADE_IN } from '@trezor/components/src/config/animations';
-import { HoverAnimation } from '../../HoverAnimation';
+
+import { HoverAnimation } from 'src/components/suite';
 
 const Wrapper = styled.div<{
     $isOpen: ActionItemProps['isOpen'];
