+++ packages/suite/src/components/suite/Preloader/SuiteLayout/NavigationBar/NavigationActions/NavSettings.tsx
