+++ packages/suite/src/components/suite/Preloader/SuiteLayout/useAnchorRemoving.tsx
